moved append button to code block functionality into service
