+++ apps/daffio/src/app/docs/guides/components/guides-content/guides-content.component.ts
@@ -4,14 +4,13 @@ import {
   input,
   effect,
   inject,
-  ComponentRef,
   ElementRef,
   afterRenderEffect,
   DestroyRef,
   ViewContainerRef,
 } from '@angular/core';
 
-import { DaffDocsCopyButtonComponent } from '@daffodil/docs';
+import { CodeBlockCopyButtonService } from '@daffodil/docs';
 import {
   DaffDoc,
   DaffDocKind,
@@ -26,6 +25,7 @@ import { DaffioDocsTocService } from '../../../toc/toc.service';
   selector: 'daffio-docs-guides-content',
   templateUrl: './guides-content.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
+  providers: [CodeBlockCopyButtonService],
   imports: [
     DaffioDocViewerComponent,
     DaffioSafeHtmlPipe,
@@ -39,13 +39,13 @@ export class DaffioDocsGuidesContentComponent implements DaffioDocsDynamicConten
   private elementRef = inject(ElementRef<HTMLElement>);
   private destroyRef = inject(DestroyRef);
   private viewContainerRef = inject(ViewContainerRef);
-  private buttonRefs: Array<ComponentRef<DaffDocsCopyButtonComponent>> = [];
 
   constructor(
     private tocRegistry: DaffioDocsTocService,
+    private copyButtonService: CodeBlockCopyButtonService,
   ) {
     this.destroyRef.onDestroy(() => {
-      this.cleanupButtons();
+      this.copyButtonService.cleanup();
     });
 
     effect((onCleanup) => {
@@ -58,54 +58,8 @@ export class DaffioDocsGuidesContentComponent implements DaffioDocsDynamicConten
     afterRenderEffect({
       write: () => {
         this.doc();
-        this.addCopyButtonsToCodeBlocks();
+        this.copyButtonService.addCopyButtonsToCodeBlocks(this.elementRef.nativeElement, this.viewContainerRef);
       },
     });
-  }
-
-  /**
-   * Finds all code blocks and adds copy buttons to them
-   */
-  private addCopyButtonsToCodeBlocks(): void {
-    this.cleanupButtons();
-
-    const codeBlocks = this.elementRef.nativeElement.querySelectorAll('pre');
-
-    codeBlocks.forEach((pre: HTMLPreElement) => {
-      const code = pre.querySelector('code');
-      if (!code) {
-        return;
-      }
-
-      const textContent = code.textContent || '';
-
-      pre.style.position = 'relative';
-
-      // Create the copy button component
-      const buttonRef = this.viewContainerRef.createComponent(DaffDocsCopyButtonComponent);
-
-      // Pass the code snippet to the button
-      buttonRef.setInput('content', textContent);
-
-      const buttonElement = buttonRef.location.nativeElement;
-
-      const innerButton = buttonElement.querySelector('button');
-      if (innerButton) {
-        innerButton.classList.add('daff-copy-code-block');
-      }
-
-      // Add copy button to the code block
-      pre.appendChild(buttonElement);
-
-      this.buttonRefs.push(buttonRef);
-    });
-  }
-
-  /**
-   * Cleanup copy button references
-   */
-  private cleanupButtons(): void {
-    this.buttonRefs.forEach(ref => ref.destroy());
-    this.buttonRefs = [];
   }
 }
