+++ libs/docs/src/copy-button/copy-button/copy-button.component.spec.ts
